+++ src/codes/pi.rs
@@ -259,64 +259,62 @@ mod tests {
     use crate::prelude::*;
 
     #[test]
-    fn test_roundtrip() -> Result<(), Box<dyn core::error::Error>> {
+    fn test_roundtrip() {
         let k = 3;
         for value in (0..64).map(|i| 1 << i).chain(0..1024).chain([u64::MAX - 1]) {
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<BE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi(value, k)?;
+            let code_len = writer.write_pi(value, k).unwrap();
             assert_eq!(code_len, len_pi(value, k));
             drop(writer);
             let mut reader = <BufBitReader<BE, _>>::new(MemWordReader::new(&data));
             assert_eq!(
-                reader.read_pi(k)?,
+                reader.read_pi(k).unwrap(),
                 value,
                 "for value: {} with k {}",
                 value,
                 k
             );
         }
-        Ok(())
     }
 
     #[test]
-    fn test_roundtrip_pi2() -> Result<(), Box<dyn core::error::Error>> {
+    fn test_roundtrip_pi2() {
         // Test the specialized pi2 methods
         for value in (0..64).map(|i| 1 << i).chain(0..1024).chain([u64::MAX - 1]) {
             // Test BE
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<BE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi2(value)?;
+            let code_len = writer.write_pi2(value).unwrap();
             assert_eq!(code_len, len_pi(value, 2));
             drop(writer);
             let mut reader = <BufBitReader<BE, _>>::new(MemWordReader::new(&data));
-            assert_eq!(reader.read_pi2()?, value, "BE for value: {}", value,);
+            assert_eq!(reader.read_pi2().unwrap(), value, "BE for value: {}", value,);
 
             // Test LE
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<LE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi2(value)?;
+            let code_len = writer.write_pi2(value).unwrap();
             assert_eq!(code_len, len_pi(value, 2));
             drop(writer);
             let mut reader = <BufBitReader<LE, _>>::new(MemWordReader::new(&data));
-            assert_eq!(reader.read_pi2()?, value, "LE for value: {}", value,);
+            assert_eq!(reader.read_pi2().unwrap(), value, "LE for value: {}", value,);
         }
-        Ok(())
     }
 
     #[test]
-    fn test_roundtrip_pi2_param() -> Result<(), Box<dyn core::error::Error>> {
+    fn test_roundtrip_pi2_param() {
         // Test the parametric pi2 methods with tables
         for value in (0..64).map(|i| 1 << i).chain(0..1024).chain([u64::MAX - 1]) {
             // Test BE with tables
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<BE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi2_param::<true>(value)?;
+            let code_len = writer.write_pi2_param::<true>(value).unwrap();
             assert_eq!(code_len, len_pi(value, 2));
             drop(writer);
             let mut reader = <BufBitReader<BE, _>>::new(MemWordReader::new(&data));
             assert_eq!(
-                reader.read_pi2_param::<true>()?,
+                reader.read_pi2_param::<true>().unwrap(),
                 value,
                 "BE table for value: {}",
                 value,
@@ -325,12 +323,12 @@ mod tests {
             // Test BE without tables
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<BE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi2_param::<false>(value)?;
+            let code_len = writer.write_pi2_param::<false>(value).unwrap();
             assert_eq!(code_len, len_pi(value, 2));
             drop(writer);
             let mut reader = <BufBitReader<BE, _>>::new(MemWordReader::new(&data));
             assert_eq!(
-                reader.read_pi2_param::<false>()?,
+                reader.read_pi2_param::<false>().unwrap(),
                 value,
                 "BE no table for value: {}",
                 value,
@@ -339,12 +337,12 @@ mod tests {
             // Test LE with tables
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<LE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi2_param::<true>(value)?;
+            let code_len = writer.write_pi2_param::<true>(value).unwrap();
             assert_eq!(code_len, len_pi(value, 2));
             drop(writer);
             let mut reader = <BufBitReader<LE, _>>::new(MemWordReader::new(&data));
             assert_eq!(
-                reader.read_pi2_param::<true>()?,
+                reader.read_pi2_param::<true>().unwrap(),
                 value,
                 "LE table for value: {}",
                 value,
@@ -353,23 +351,22 @@ mod tests {
             // Test LE without tables
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<LE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi2_param::<false>(value)?;
+            let code_len = writer.write_pi2_param::<false>(value).unwrap();
             assert_eq!(code_len, len_pi(value, 2));
             drop(writer);
             let mut reader = <BufBitReader<LE, _>>::new(MemWordReader::new(&data));
             assert_eq!(
-                reader.read_pi2_param::<false>()?,
+                reader.read_pi2_param::<false>().unwrap(),
                 value,
                 "LE no table for value: {}",
                 value,
             );
         }
-        Ok(())
     }
 
     #[test]
     #[allow(clippy::unusual_byte_groupings)]
-    fn test_bits() -> Result<(), Box<dyn core::error::Error>> {
+    fn test_bits() {
         for (k, value, expected) in [
             (2, 20, 0b01_00_0101 << (64 - 8)),
             (2, 0, 0b100 << (64 - 3)),
@@ -391,7 +388,7 @@ mod tests {
         ] {
             let mut data = [0_u64; 10];
             let mut writer = <BufBitWriter<BE, _>>::new(MemWordWriterSlice::new(&mut data));
-            let code_len = writer.write_pi(value, k)?;
+            let code_len = writer.write_pi(value, k).unwrap();
             assert_eq!(code_len, len_pi(value, k));
             drop(writer);
             assert_eq!(
@@ -406,11 +403,10 @@ mod tests {
                 len_pi(value, k),
             );
         }
-        Ok(())
     }
 
     #[test]
-    fn test_against_zeta() -> Result<(), Box<dyn core::error::Error>> {
+    fn test_against_zeta() {
         // BE: π₀ = ζ₁ and π₁ = ζ₂
         for k in 0..2 {
             for value in 0..100 {
@@ -418,13 +414,13 @@ mod tests {
                 let mut data_zeta = [0_u64; 10];
 
                 let mut writer = <BufBitWriter<BE, _>>::new(MemWordWriterSlice::new(&mut data_pi));
-                let code_len = writer.write_pi(value, k)?;
+                let code_len = writer.write_pi(value, k).unwrap();
                 assert_eq!(code_len, len_pi(value, k));
                 drop(writer);
 
                 let mut writer =
                     <BufBitWriter<BE, _>>::new(MemWordWriterSlice::new(&mut data_zeta));
-                let code_len = writer.write_zeta(value, 1 << k)?;
+                let code_len = writer.write_zeta(value, 1 << k).unwrap();
                 assert_eq!(code_len, len_zeta(value, 1 << k));
                 drop(writer);
 
@@ -438,12 +434,12 @@ mod tests {
             let mut data_zeta = [0_u64; 10];
 
             let mut writer = <BufBitWriter<LE, _>>::new(MemWordWriterSlice::new(&mut data_pi));
-            let code_len = writer.write_pi(value, 0)?;
+            let code_len = writer.write_pi(value, 0).unwrap();
             assert_eq!(code_len, len_pi(value, 0));
             drop(writer);
 
             let mut writer = <BufBitWriter<LE, _>>::new(MemWordWriterSlice::new(&mut data_zeta));
-            let code_len = writer.write_zeta(value, 1)?;
+            let code_len = writer.write_zeta(value, 1).unwrap();
             assert_eq!(code_len, len_zeta(value, 1));
             drop(writer);
 
@@ -453,6 +449,5 @@ mod tests {
         for value in 0..100 {
             assert_eq!(len_pi(value, 1), len_zeta(value, 2));
         }
-        Ok(())
     }
 }