+++ src/impls/mem_word_reader.rs
@@ -26,7 +26,7 @@ use mem_dbg::{MemDbg, MemSize};
 /// # Examples
 ///
 /// ```rust
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+/// # fn main() -> Result<(), Box<dyn core::error::Error>> {
 /// use dsi_bitstream::prelude::*;
 ///
 /// let words: [u32; 2] = [1, 2];
@@ -155,23 +155,22 @@ impl<W: Word, B: AsRef<[W]>> WordSeek for MemWordReader<W, B, false> {
 }
 
 #[test]
-fn test_eof_table_read() -> Result<(), Box<dyn core::error::Error>> {
+fn test_eof_table_read() {
     use crate::codes::{DeltaReadParam, DeltaWrite};
     let mut words: [u64; 1] = [0];
     let mut writer = crate::prelude::BufBitWriter::<crate::prelude::LE, _>::new(
         MemWordWriterSlice::new(&mut words),
     );
     for _ in 0..16 {
-        writer.write_delta(1)?;
+        writer.write_delta(1).unwrap();
     }
-    writer.flush()?;
+    writer.flush().unwrap();
     drop(writer);
 
     let mut reader =
         crate::prelude::BufBitReader::<crate::prelude::LE, _>::new(MemWordReader::new(&words));
     for _ in 0..16 {
         // Here the last table read returns zero-extended data
-        assert_eq!(1, reader.read_delta_param::<true, true>()?);
+        assert_eq!(1, reader.read_delta_param::<true, true>().unwrap());
     }
-    Ok(())
 }